requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5

# Production server
gunicorn==21.2.0

# File processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Environment and configuration
python-dotenv==1.0.0
cfenv==0.5.3

# API and HTTP
requests==2.31.0
urllib3==2.0.7

# Rate limiting and security
Flask-Limiter==3.5.0
PyJWT==2.8.0

# System monitoring
psutil==5.9.6

# Date and time handling
python-dateutil==2.8.2

# JSON handling (built-in, but explicit version for containers)
# json - built-in modulePyJWT==2.8.0
cryptography==41.0.7
